Add new file for supporting decParts arithmetic functions
