+++ decimal64math.go
@@ -288,51 +288,3 @@ func (ctx Context64) Mul(d, e Decimal64) Decimal64 {
 	}
 	return newFromParts(ans.sign, ans.exp, ans.significand.lo)
 }
-
-func (dp *decParts) roundToLo() discardedDigit {
-	var rndStatus discardedDigit
-	if dp.significand.numDecimalDigits() > 16 {
-		var remainder uint64
-		expDiff := dp.significand.numDecimalDigits() - 16
-		dp.exp += expDiff
-		dp.significand, remainder = dp.significand.divrem64(powersOf10[expDiff])
-		rndStatus = roundStatus(remainder, 0, expDiff)
-	}
-	return rndStatus
-}
-
-func (dp *decParts) matchScales128(ep *decParts) {
-	targetExp := ep.exp - dp.exp
-	if (ep.significand != uint128T{0, 0}) {
-		if targetExp < 0 {
-			dp.significand = dp.significand.mul(powerOfTen128(targetExp))
-			dp.exp += targetExp
-		} else if targetExp > 0 {
-			ep.significand = ep.significand.mul(powerOfTen128(targetExp))
-			ep.exp -= (targetExp)
-		}
-	}
-	return
-}
-
-// add128 adds two decParts with full precision in 128 bits of significand
-func (dp *decParts) add128(ep *decParts) decParts {
-	logicCheck(ep.exp == dp.exp, "ep.exp == dp.exp")
-	var ans decParts
-	ans.exp = dp.exp
-	if dp.sign == ep.sign {
-		ans.sign = dp.sign
-		ans.significand = dp.significand.add(ep.significand)
-	} else {
-		if ep.significand.gt(dp.significand) {
-			ans.sign = ep.sign
-			ans.significand = ep.significand.sub(dp.significand)
-		} else if ep.significand.lt(dp.significand) {
-			ans.sign = dp.sign
-			ans.significand = dp.significand.sub(ep.significand)
-		} else {
-			ans.significand = uint128T{0, 0}
-		}
-	}
-	return ans
-}
